Cambios finales

+++ QuevedoVirtualClassrooms-Server/src/main/java/com/quevedo/virtualclassroomsserver/logic/dao/ResourceDaoImpl.java
@@ -6,17 +6,27 @@ import com.quevedo.virtualclassroomsserver.common.models.common.ResourceType;
 import com.quevedo.virtualclassroomsserver.common.models.dto.resource.ResourceDetailGetDTO;
 import com.quevedo.virtualclassroomsserver.common.models.dto.resource.ResourceLiteGetDTO;
 import com.quevedo.virtualclassroomsserver.common.models.dto.resource.ResourcePostDTO;
+import com.quevedo.virtualclassroomsserver.common.models.dto.resource.ResourcePutDTO;
+import com.quevedo.virtualclassroomsserver.common.models.dto.resource.comment.ResourceCommentPost;
 import com.quevedo.virtualclassroomsserver.common.models.server.resource.Resource;
 import com.quevedo.virtualclassroomsserver.common.models.server.resource.ResourceComment;
+import com.quevedo.virtualclassroomsserver.common.models.server.user.UserVisualization;
 import com.quevedo.virtualclassroomsserver.facade.dao.ResourceDao;
 import com.quevedo.virtualclassroomsserver.logic.dao.utils.rowmappers.CommentRowMapper;
 import com.quevedo.virtualclassroomsserver.logic.dao.utils.rowmappers.ResourceRowMapper;
+import com.quevedo.virtualclassroomsserver.logic.dao.utils.rowmappers.VisualizationRowMapper;
 import io.vavr.control.Either;
 import jakarta.inject.Inject;
+import jakarta.ws.rs.core.Context;
+import jakarta.ws.rs.core.SecurityContext;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.dao.DataAccessException;
 import org.springframework.dao.IncorrectResultSizeDataAccessException;
+import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.PreparedStatementSetter;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
 
 import java.io.File;
 import java.io.IOException;
@@ -24,8 +34,11 @@ import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.security.Principal;
 import java.sql.PreparedStatement;
+import java.sql.Statement;
 import java.sql.Timestamp;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
@@ -37,18 +50,21 @@ public class ResourceDaoImpl implements ResourceDao {
     private final QueriesLoader queriesLoader;
     private final ResourceRowMapper resourceRowMapper;
     private final CommentRowMapper commentRowMapper;
+    private final VisualizationRowMapper userVisualizationRowMapper;
 
     @Inject
     public ResourceDaoImpl(JdbcTemplate template,
                            Config config,
                            QueriesLoader queriesLoader,
                            ResourceRowMapper resourceRowMapper,
-                           CommentRowMapper commentRowMapper) {
+                           CommentRowMapper commentRowMapper,
+                           VisualizationRowMapper userVisualizationRowMapper) {
         this.commentRowMapper = commentRowMapper;
         this.resourceRowMapper = resourceRowMapper;
         this.template = template;
         this.config = config;
         this.queriesLoader = queriesLoader;
+        this.userVisualizationRowMapper = userVisualizationRowMapper;
     }
 
 
@@ -113,11 +129,10 @@ public class ResourceDaoImpl implements ResourceDao {
     }
 
     @Override
-    public Either<String, Integer> editResource(ResourcePostDTO toEdit) {
-        //TODO: verificar si queremos editar recursos y crear modelo para hacerlo
+    public Either<String, Integer> editResource(ResourcePutDTO toEdit) {
         Either<String, Integer> result;
         try {
-            result = Either.right(template.update(queriesLoader.getUpdateResource(), toEdit.getResourceName(), toEdit.getClassroomUUID(), ""));
+            result = Either.right(template.update(queriesLoader.getUpdateResource(), toEdit.getResourceName(), toEdit.getClassroomUUID(), toEdit.getResourceType().getValue(),toEdit.getUuidResource()));
         }catch (DataAccessException dataAccessException){
             result = Either.left("ERROR: " + dataAccessException);
         }
@@ -125,10 +140,14 @@ public class ResourceDaoImpl implements ResourceDao {
     }
 
     @Override
-    public Either<String, List<ResourceLiteGetDTO>> getAllResourcesOfClassroom(String classroomId, ResourceType resourceType) {
+    public Either<String, List<ResourceLiteGetDTO>> getAllResourcesOfClassroom(String classroomId, String resourceType) {
         Either<String, List<ResourceLiteGetDTO>> result;
+        if (!resourceType.equals("%%")){
+            ResourceType resourceType1 = ResourceType.valueOf(resourceType);
+            resourceType = Integer.toString(resourceType1.getValue());
+        }
         try {
-            result = Either.right(template.query(queriesLoader.getGetAllResourcesByClassroomId(), resourceRowMapper, classroomId, resourceType.getValue()).stream().map(Resource::toLiteDTO).collect(Collectors.toList()));
+            result = Either.right(template.query(queriesLoader.getGetAllResourcesByClassroomId(), resourceRowMapper, classroomId, resourceType).stream().map(Resource::toLiteDTO).collect(Collectors.toList()));
         } catch (DataAccessException dataAccessException) {
             result = Either.left("ERROR: " + dataAccessException);
         }
@@ -140,7 +159,10 @@ public class ResourceDaoImpl implements ResourceDao {
         Either<String, ResourceDetailGetDTO> result;
         try {
             Resource resource = template.queryForObject(queriesLoader.getGetResourceById(), resourceRowMapper, resourceUUID);
-            Objects.requireNonNull(resource).setComments(prepareList(template.query(queriesLoader.getGetAllCommentsByResource(), commentRowMapper, resourceUUID)));
+            List<ResourceComment> comments = template.query(queriesLoader.getGetAllCommentsByResource(), commentRowMapper, resourceUUID);
+            List<UserVisualization> visualizations = template.query(queriesLoader.getGetVisualizationByResource(), userVisualizationRowMapper, resourceUUID);
+            resource.setComments(comments);
+            resource.setVisualizations(visualizations);
             result = Either.right(resource.toDetailDTO());
 
         } catch (DataAccessException dataAccessException) {
@@ -152,7 +174,7 @@ public class ResourceDaoImpl implements ResourceDao {
     }
 
     @Override
-    public Either<String, File> getResourceFile(String fileName) {
+    public Either<String, File> getResourceFile(String fileName, String username) {
         Either<String, File> result;
         String resourceUUID = fileName.split("\\.")[0];
         try {
@@ -169,12 +191,43 @@ public class ResourceDaoImpl implements ResourceDao {
                     return Either.left("Not supported resource, yet...");
             }
             if (file.exists()) {
+                template.update(queriesLoader.getInsertVisualization(), username, resourceUUID, Timestamp.valueOf(LocalDateTime.now()));
                 result = Either.right(file);
             } else {
                 result = Either.left("The file of this resource is not at the directory, contact admin");
             }
         } catch (IncorrectResultSizeDataAccessException sizeException) {
             result = Either.left("This resource does not exist");
+        }
+        return result;
+    }
+
+    @Override
+    public Either<String, ResourceComment> postNewComment(ResourceCommentPost comment) {
+        Either<String, ResourceComment> result;
+        ResourceComment resourceComment = ResourceComment.builder()
+                .text(comment.getText())
+                .usernameOwner(comment.getUsername())
+                .timeStamp(LocalDateTime.now())
+                .answersTo(comment.getAnswersTo())
+                .build();
+        try {
+            KeyHolder keyHolder = new GeneratedKeyHolder();
+            template.update(connection -> {
+                PreparedStatement ps = connection
+                        .prepareStatement(queriesLoader.getInsertComment(), Statement.RETURN_GENERATED_KEYS);
+                ps.setString(1, resourceComment.getText());
+                ps.setTimestamp(2, Timestamp.valueOf(resourceComment.getTimeStamp()));
+                ps.setInt(3, resourceComment.getAnswersTo());
+                ps.setString(4, resourceComment.getUsernameOwner());
+                ps.setString(5, comment.getResourceId());
+                return ps;
+            }, keyHolder);
+            resourceComment.setIdComment(keyHolder.getKey().intValue());
+            result = Either.right(resourceComment);
+        }catch (DataAccessException dataAccessException){
+            log.error(dataAccessException.getMessage(), dataAccessException);
+            result = Either.left(dataAccessException.getMessage());
         }
         return result;
     }
@@ -196,38 +249,25 @@ public class ResourceDaoImpl implements ResourceDao {
                     return "Not supported resource, yet...";
             }
             if (template.update(queriesLoader.getDeleteResourceById(), resourceUUID) > 0){
-                errorResult.append(" Deleted resource");
+                errorResult.append(" Deleted resource reference");
             }else {
                 errorResult.append(" The record of this resource could not be deleted");
             }
-            if (file.exists()){
-                if (file.delete()){
-                    errorResult.append("Deleted File");
+            if (resource.getResourceType() != ResourceType.URL){
+                if (file.exists()){
+                    if (file.delete()){
+                        errorResult.append(" Deleted File");
+                    }else {
+                        errorResult.append(" The file of this resource exists but could not be deleted");
+                    }
                 }else {
-                    errorResult.append("The file of this resource exists but could not be deleted");
+                    errorResult.append(" The file of this resource could not be found.");
                 }
-            }else {
-                errorResult.append("The file of this resource could not be found.");
             }
             return errorResult.toString();
         }catch (DataAccessException dataAccessException){
             errorResult = new StringBuilder("ERROR: " + dataAccessException);
         }
         return errorResult.toString();
-    }
-
-    private List<ResourceComment> prepareList(List<ResourceComment> input) {
-        List<ResourceComment> notAnswers = input.stream().filter(comment -> !comment.isAnswer()).collect(Collectors.toList());
-        List<ResourceComment> answers = input.stream().filter(ResourceComment::isAnswer).collect(Collectors.toList());
-
-        return notAnswers.stream()
-                .map(comment -> {
-                    answers.forEach(answer -> {
-                        if (comment.getUuidComment().toString().equals(answer.getAnswersTo().toString())) {
-                            comment.pushCommentToAnswers(answer);
-                        }
-                    });
-                    return comment;
-                }).collect(Collectors.toList());
     }
 }
